// experiments/ax2/runs/gpt-openai/C2-expert/rep1/project/src/comments/comment.types.ts
import { ProfileResponse } from "../profiles/profile.types";

export interface CommentRecord {
  readonly id: number;
  readonly body: string;
  readonly createdAt: Date;
  readonly updatedAt: Date;
  readonly authorId: number;
  readonly articleId: number;
  readonly author: {
    readonly username: string;
    readonly bio: string | null;
    readonly image: string | null;
  };
}

export interface CommentResponse {
  readonly id: number;
  readonly createdAt: string;
  readonly updatedAt: string;
  readonly body: string;
  readonly author: ProfileResponse;
}